feat(InputFile): migrate to Tailwind
